api: reorganize some header includes
There is no reason to include plugin.hpp in all core files. Some plugins
were missing it.
