+++ plugins/single_plugins/move.cpp
@@ -1,3 +1,4 @@
+#include <plugin.hpp>
 #include <output.hpp>
 #include <core.hpp>
 #include <view.hpp>
