+++ plugins/single_plugins/resize.cpp
@@ -1,3 +1,4 @@
+#include <plugin.hpp>
 #include <output.hpp>
 #include <view.hpp>
 #include <core.hpp>
