+++ plugins/single_plugins/vswitch.cpp
@@ -1,3 +1,4 @@
+#include <plugin.hpp>
 #include <output.hpp>
 #include <core.hpp>
 #include <debug.hpp>
