+++ src/core/img.cpp
@@ -11,7 +11,6 @@
 #include <stdint.h>
 #include <unistd.h>
 #include <cstdio>
-#include <iostream>
 #include <unordered_map>
 #include <functional>
