+++ src/core/wm.hpp
@@ -1,6 +1,8 @@
 #ifndef WM_H
 #define WM_H
+
 #include "plugin.hpp"
+#include "bindings.hpp"
 #include "view.hpp"
 
 class wayfire_close : public wayfire_plugin_t {
